+++ compiler-utils/src/main/kotlin/com/r0adkll/kimchi/util/ksp/KspUtil.kt
@@ -6,6 +6,7 @@ import com.google.devtools.ksp.symbol.ClassKind
 import com.google.devtools.ksp.symbol.KSAnnotated
 import com.google.devtools.ksp.symbol.KSAnnotation
 import com.google.devtools.ksp.symbol.KSClassDeclaration
+import com.google.devtools.ksp.symbol.KSType
 import com.google.devtools.ksp.symbol.KSTypeAlias
 import com.google.devtools.ksp.symbol.KSTypeReference
 import com.squareup.kotlinpoet.ClassName
@@ -62,6 +63,9 @@ public fun KSTypeReference.findActualType(): KSClassDeclaration {
     resolvedType as KSClassDeclaration
   }
 }
+
+public val KSType.classDeclaration: KSClassDeclaration
+  get() = declaration as KSClassDeclaration
 
 public val KSClassDeclaration.isInterface: Boolean
   get() = this.classKind == ClassKind.INTERFACE